# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: PicoClaw API
  version: "1.0.0"
  description: |
    HTTP API for the PicoClaw launcher backend (port 18800) and gateway health server (port 18790).

    ## Authentication

    All `/api/*` routes require authentication except the public auth endpoints listed below.
    Two credential types are accepted interchangeably:

    - **Session cookie**: `picoclaw_launcher_auth=<value>` — set by `POST /api/auth/login`, valid for 7 days
    - **Bearer token**: `Authorization: Bearer <dashboard-token>` — plaintext dashboard token from
      `PICOCLAW_LAUNCHER_TOKEN` env var or launcher config

    ### Public endpoints (no auth required)
    - `POST /api/auth/login`
    - `POST /api/auth/logout`
    - `GET /api/auth/status`
    - `POST /api/auth/setup` (open when no password set; requires session when password already exists)

    ### WebSocket proxy (`GET /pico/ws`)
    Authenticates via the `Sec-Websocket-Protocol` header rather than a session cookie or
    bearer token. Set the header to `token.<pico-token>` during the WebSocket upgrade handshake.
    The token is obtained from `GET /api/pico/token`.
servers:
  - url: http://localhost:18800
    description: Launcher backend — web console and management API
  - url: http://localhost:18790
    description: Gateway health server — used in backend-only (picoclaw-gateway) mode
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/webhook/process:
    post:
      tags: [webhook]
      summary: Submit an asynchronous processing job with webhook callback
      description: |
        Accepts a processing job and returns immediately with a job ID (202 Accepted).
        The job runs in the background, and results are POSTed to the provided webhook URL
        when processing completes.

        **Processing flow:**
        1. Submit job → receive job_id
        2. Backend processes asynchronously
        3. Result POSTed to webhook_url

        **Webhook callback payload (success):**
        ```json
        {
          "job_id": "uuid",
          "status": "completed",
          "result": { "processed_data": "..." },
          "timestamp": "2026-04-17T10:00:05Z"
        }
        ```

        **Webhook callback payload (error):**
        ```json
        {
          "job_id": "uuid",
          "status": "failed",
          "error": "error message",
          "timestamp": "2026-04-17T10:00:05Z"
        }
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookProcessRequest"
      responses:
        "202":
          description: Job accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookProcessResponse"
        "400":
          description: Invalid request body or missing webhook_url
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/webhook/status:
    get:
      tags: [webhook]
      summary: Check the status of a submitted job
      description: |
        Query the current status of a webhook processing job.
        The job remains in memory for 2 hours after submission.
      parameters:
        - name: job_id
          in: query
          required: true
          description: Job UUID returned from POST /api/webhook/process
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookJobStatus"
        "400":
          description: Missing job_id query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Job not found (may have been cleaned up)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── GATEWAY HEALTH ───────────────────────────────────────────────────────
...
  /reload:
    post:
      tags: [gateway-health]
      summary: Trigger a config reload on the gateway
      description: |
        If the gateway was started with an auth token, include `Authorization: Bearer <token>`.
        If no token is configured, the endpoint accepts unauthenticated requests.
      security:
        - bearerAuth: []
        - {}
      servers:
        - url: http://localhost:18790
          description: Gateway health server
      responses:
        "200":
          description: Reload triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: reload triggered
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Reload handler not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: picoclaw_launcher_auth
      description: Session cookie set after successful login via POST /api/auth/login. Valid for 7 days.
    bearerAuth:
      type: http
      scheme: bearer
      description: Plaintext dashboard token from PICOCLAW_LAUNCHER_TOKEN env var or launcher config.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: unauthorized
...
    WebhookProcessRequest:
      type: object
      required: [webhook_url]
      properties:
        webhook_url:
          type: string
          format: uri
          description: URL where results will be POSTed when processing completes
          example: "https://your-app.com/webhook/callback"
        payload:
          type: object
          description: Arbitrary JSON payload to be processed
          additionalProperties: true
          example:
            data: "your data here"
            priority: "high"
    WebhookProcessResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier for this job
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          enum: [processing]
          description: Initial status is always "processing"
          example: processing
        timestamp:
          type: string
          format: date-time
          description: Job submission timestamp
          example: "2026-04-17T10:00:00Z"
    WebhookJobStatus:
      type: object
      properties:
        ID:
          type: string
          format: uuid
          description: Job identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        WebhookURL:
          type: string
          description: Callback URL for this job
          example: "https://your-app.com/webhook/callback"
        Payload:
          type: object
          description: Original payload submitted with the job
          additionalProperties: true
        Status:
          type: string
          enum: [processing, completed, failed]
          description: Current job status
          example: completed
        CreatedAt:
          type: string
          format: date-time
          description: Job creation timestamp
          example: "2026-04-17T10:00:00Z"
        CompletedAt:
          type: string
          format: date-time
          nullable: true
          description: Job completion timestamp (null if still processing)
          example: "2026-04-17T10:00:05Z"